web client centralizado
